refactor: Move all helper function to utils for consistency
