+++ src/compiler/compiler.mjs
@@ -1,4 +1,4 @@
-import { expandPositionWithLayout, inferLayoutFromKeywords } from './positionUtils.mjs';
+import { expandPositionWithLayout, inferLayoutFromKeywords } from '../utils/positionUtils.mjs';
 import { isMethodSupported } from '../components/languageConfig.js';
 import { getMethodNameFromCommand } from '../parser/reconstructor.mjs';
 
@@ -9,7 +9,7 @@ import { generateTree } from "./types/generateTree.mjs";
 import { generateMatrix } from "./types/generateMatrix.mjs";
 import { generateGraph } from "./types/generateGraph.mjs";
 import { generateText } from "./types/generateText.mjs";
-import { getMermaidContainerSize } from "./positionUtils.mjs";
+import { getMermaidContainerSize } from "../utils/positionUtils.mjs";
 
 // Helper function to generate a new node name for component types that use nodes
 function generateNodeName(body, componentType) {
