+++ src/compiler/types/generateStack.mjs
@@ -1,5 +1,5 @@
 import { formatNodeName, formatNullValue } from '../compiler.mjs';
-import { formatPositionForOutput } from '../positionUtils.mjs';
+import { formatPositionForOutput } from '../../utils/positionUtils.mjs';
 
 export function generateStack(stackComponent, layout = [3, 3]) {
     let result = "stack\n";
